Update Quarkus to  3.34.2 and quarkus-github-app to 2.14.0
