+++ .github/workflows/polling-tests.yml
@@ -26,4 +26,4 @@ jobs:
         java-version: ${{ matrix.java-version }}
 
     - name: Build with Maven
-      run: ./mvnw -B test -Dquarkus.test.profile=test,polling
+      run: ./mvnw -B test -Dquarkus.test.profile=polling,test # order is important, test profile should be last to override polling's configuration
